Сетевые запросы переписаны для взаимодействия с собственным прокси-сервером
