+++ attestation1/delta/src/actions/authorization.js
@@ -1,4 +1,4 @@
-import { getUserDataApi } from "../api/dummyApi";
+import { getUserDataApi } from "../api/dummyProxyApi";
 
 const authUserAction = (user) => ({
     type: 'AUTH_FORM/AUTHORIZE_USER',
