+++ attestation1/delta/src/actions/loadComments.js
@@ -1,4 +1,4 @@
-import { getCommentsApi } from "../api/dummyApi";
+import { getCommentsApi } from "../api/dummyProxyApi";
 
 const showLoadingAction = () => ({
     type: 'OPEN_POST/SHOW_LOADING',
