--- attestation1/delta/src/constants/dummyApi.js
@@ -1,9 +1,0 @@
-const BASE_URL = 'https://dummyapi.io/data/v1/';
-
-export const USER_LIST_URL = `${BASE_URL}user`;
-
-export const APP_ID_VALUE = '617d0f3dab5df4b9bd1507dd';
-
-export const APP_ID_FIELD = 'app-id';
-
-export const POST_LIST_URL = `${BASE_URL}post`;
