+++ attestation1/delta/src/constants/dummyProxyApi.js
@@ -1,0 +1,13 @@
+const BASE_URL = 'http://127.0.0.1:3000/';
+
+export const USER_LIST_URL = `${BASE_URL}user`;
+
+export const POST_LIST_URL = `${BASE_URL}post`;
+
+export const COMMENT_LIST_URL = `${BASE_URL}comment`;
+
+export const GET_METHOD = 'GET';
+
+export const POST_METHOD = 'POST';
+
+export const PUT_METHOD = 'PUT';
